requirements.txt
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bidict==0.22.1
bleach==6.0.0
blinker==1.6.2
cachelib==0.10.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.7
comm==0.1.4
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.1
entrypoints==0.4
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Session==0.5.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
fqdn==1.5.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
inotify==0.2.10
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter_client==7.4.9
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.1
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
nose==1.3.7
notebook==6.5.5
notebook_shim==0.2.3
numpy==1.25.2
overrides==7.4.0
packaging==23.1
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
plotly==5.7.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pymongo==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-json-logger==2.0.7
python-socketio==5.10.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.4.4
QtPy==2.4.0
redis==5.0.0
referencing==0.30.2
requests==2.27.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.2
Send2Trash==1.8.2
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.20
stack-data==0.6.2
tailer==0.4.1
tenacity==8.2.3
terminado==0.17.1
tinycss2==1.2.1
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.7.1
uri-template==1.3.0
urllib3==1.26.16
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
Werkzeug==2.3.7
widgetsnbextension==4.0.8
wsproto==1.2.0
